DEVOPS-388: Remove file path escaping.
It does not seem to work correctly, i. e. Resource Hacker does not seem
to handle the escape sequences as it should.

We now only surround paths with double quotes; since double quotes are
not valid in the Win32 file namespace, we can probably get away with NOT
escaping double quotes inside filenames.